# kgs-platform/api/registry/v1/registry_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v4.25.1
// source: registry/v1/registry.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationRegistryCreateApp = "/api.registry.v1.Registry/CreateApp"
const OperationRegistryGetApp = "/api.registry.v1.Registry/GetApp"
const OperationRegistryIssueApiKey = "/api.registry.v1.Registry/IssueApiKey"
const OperationRegistryListApps = "/api.registry.v1.Registry/ListApps"
const OperationRegistryRevokeApiKey = "/api.registry.v1.Registry/RevokeApiKey"

type RegistryHTTPServer interface {
	// CreateApp Create a new application
	CreateApp(context.Context, *CreateAppRequest) (*CreateAppReply, error)
	// GetApp Get application details
	GetApp(context.Context, *GetAppRequest) (*GetAppReply, error)
	// IssueApiKey Issue a new API key for an application
	IssueApiKey(context.Context, *IssueApiKeyRequest) (*IssueApiKeyReply, error)
	// ListApps List all applications
	ListApps(context.Context, *ListAppsRequest) (*ListAppsReply, error)
	// RevokeApiKey Revoke an existing API key
	RevokeApiKey(context.Context, *RevokeApiKeyRequest) (*RevokeApiKeyReply, error)
}

func RegisterRegistryHTTPServer(s *http.Server, srv RegistryHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/apps", _Registry_CreateApp0_HTTP_Handler(srv))
	r.GET("/v1/apps/{app_id}", _Registry_GetApp0_HTTP_Handler(srv))
	r.GET("/v1/apps", _Registry_ListApps0_HTTP_Handler(srv))
	r.POST("/v1/apps/{app_id}/keys", _Registry_IssueApiKey0_HTTP_Handler(srv))
	r.DELETE("/v1/keys/{key_hash}", _Registry_RevokeApiKey0_HTTP_Handler(srv))
}

func _Registry_CreateApp0_HTTP_Handler(srv RegistryHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateAppRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRegistryCreateApp)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateApp(ctx, req.(*CreateAppRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateAppReply)
		return ctx.Result(200, reply)
	}
}

func _Registry_GetApp0_HTTP_Handler(srv RegistryHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetAppRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRegistryGetApp)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetApp(ctx, req.(*GetAppRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetAppReply)
		return ctx.Result(200, reply)
	}
}

func _Registry_ListApps0_HTTP_Handler(srv RegistryHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListAppsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRegistryListApps)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListApps(ctx, req.(*ListAppsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListAppsReply)
		return ctx.Result(200, reply)
	}
}

func _Registry_IssueApiKey0_HTTP_Handler(srv RegistryHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in IssueApiKeyRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRegistryIssueApiKey)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.IssueApiKey(ctx, req.(*IssueApiKeyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*IssueApiKeyReply)
		return ctx.Result(200, reply)
	}
}

func _Registry_RevokeApiKey0_HTTP_Handler(srv RegistryHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RevokeApiKeyRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRegistryRevokeApiKey)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RevokeApiKey(ctx, req.(*RevokeApiKeyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RevokeApiKeyReply)
		return ctx.Result(200, reply)
	}
}

type RegistryHTTPClient interface {
	// CreateApp Create a new application
	CreateApp(ctx context.Context, req *CreateAppRequest, opts ...http.CallOption) (rsp *CreateAppReply, err error)
	// GetApp Get application details
	GetApp(ctx context.Context, req *GetAppRequest, opts ...http.CallOption) (rsp *GetAppReply, err error)
	// IssueApiKey Issue a new API key for an application
	IssueApiKey(ctx context.Context, req *IssueApiKeyRequest, opts ...http.CallOption) (rsp *IssueApiKeyReply, err error)
	// ListApps List all applications
	ListApps(ctx context.Context, req *ListAppsRequest, opts ...http.CallOption) (rsp *ListAppsReply, err error)
	// RevokeApiKey Revoke an existing API key
	RevokeApiKey(ctx context.Context, req *RevokeApiKeyRequest, opts ...http.CallOption) (rsp *RevokeApiKeyReply, err error)
}

type RegistryHTTPClientImpl struct {
	cc *http.Client
}

func NewRegistryHTTPClient(client *http.Client) RegistryHTTPClient {
	return &RegistryHTTPClientImpl{client}
}

// CreateApp Create a new application
func (c *RegistryHTTPClientImpl) CreateApp(ctx context.Context, in *CreateAppRequest, opts ...http.CallOption) (*CreateAppReply, error) {
	var out CreateAppReply
	pattern := "/v1/apps"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRegistryCreateApp))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetApp Get application details
func (c *RegistryHTTPClientImpl) GetApp(ctx context.Context, in *GetAppRequest, opts ...http.CallOption) (*GetAppReply, error) {
	var out GetAppReply
	pattern := "/v1/apps/{app_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRegistryGetApp))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// IssueApiKey Issue a new API key for an application
func (c *RegistryHTTPClientImpl) IssueApiKey(ctx context.Context, in *IssueApiKeyRequest, opts ...http.CallOption) (*IssueApiKeyReply, error) {
	var out IssueApiKeyReply
	pattern := "/v1/apps/{app_id}/keys"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRegistryIssueApiKey))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListApps List all applications
func (c *RegistryHTTPClientImpl) ListApps(ctx context.Context, in *ListAppsRequest, opts ...http.CallOption) (*ListAppsReply, error) {
	var out ListAppsReply
	pattern := "/v1/apps"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRegistryListApps))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// RevokeApiKey Revoke an existing API key
func (c *RegistryHTTPClientImpl) RevokeApiKey(ctx context.Context, in *RevokeApiKeyRequest, opts ...http.CallOption) (*RevokeApiKeyReply, error) {
	var out RevokeApiKeyReply
	pattern := "/v1/keys/{key_hash}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRegistryRevokeApiKey))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}