# kgs-platform/api/rules/v1/rules_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v4.25.1
// source: rules/v1/rules.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationRulesCreateRule = "/api.rules.v1.Rules/CreateRule"
const OperationRulesGetRule = "/api.rules.v1.Rules/GetRule"
const OperationRulesListRules = "/api.rules.v1.Rules/ListRules"

type RulesHTTPServer interface {
	// CreateRule Create a new business rule
	CreateRule(context.Context, *CreateRuleRequest) (*RuleReply, error)
	// GetRule Get a specific rule
	GetRule(context.Context, *GetRuleRequest) (*RuleReply, error)
	// ListRules List all rules for the app
	ListRules(context.Context, *ListRulesRequest) (*ListRulesReply, error)
}

func RegisterRulesHTTPServer(s *http.Server, srv RulesHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/rules", _Rules_CreateRule0_HTTP_Handler(srv))
	r.GET("/v1/rules", _Rules_ListRules0_HTTP_Handler(srv))
	r.GET("/v1/rules/{id}", _Rules_GetRule0_HTTP_Handler(srv))
}

func _Rules_CreateRule0_HTTP_Handler(srv RulesHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateRuleRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRulesCreateRule)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateRule(ctx, req.(*CreateRuleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RuleReply)
		return ctx.Result(200, reply)
	}
}

func _Rules_ListRules0_HTTP_Handler(srv RulesHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListRulesRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRulesListRules)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListRules(ctx, req.(*ListRulesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListRulesReply)
		return ctx.Result(200, reply)
	}
}

func _Rules_GetRule0_HTTP_Handler(srv RulesHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetRuleRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRulesGetRule)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetRule(ctx, req.(*GetRuleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RuleReply)
		return ctx.Result(200, reply)
	}
}

type RulesHTTPClient interface {
	// CreateRule Create a new business rule
	CreateRule(ctx context.Context, req *CreateRuleRequest, opts ...http.CallOption) (rsp *RuleReply, err error)
	// GetRule Get a specific rule
	GetRule(ctx context.Context, req *GetRuleRequest, opts ...http.CallOption) (rsp *RuleReply, err error)
	// ListRules List all rules for the app
	ListRules(ctx context.Context, req *ListRulesRequest, opts ...http.CallOption) (rsp *ListRulesReply, err error)
}

type RulesHTTPClientImpl struct {
	cc *http.Client
}

func NewRulesHTTPClient(client *http.Client) RulesHTTPClient {
	return &RulesHTTPClientImpl{client}
}

// CreateRule Create a new business rule
func (c *RulesHTTPClientImpl) CreateRule(ctx context.Context, in *CreateRuleRequest, opts ...http.CallOption) (*RuleReply, error) {
	var out RuleReply
	pattern := "/v1/rules"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRulesCreateRule))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetRule Get a specific rule
func (c *RulesHTTPClientImpl) GetRule(ctx context.Context, in *GetRuleRequest, opts ...http.CallOption) (*RuleReply, error) {
	var out RuleReply
	pattern := "/v1/rules/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRulesGetRule))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListRules List all rules for the app
func (c *RulesHTTPClientImpl) ListRules(ctx context.Context, in *ListRulesRequest, opts ...http.CallOption) (*ListRulesReply, error) {
	var out ListRulesReply
	pattern := "/v1/rules"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRulesListRules))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}