# kgs-platform/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/apps:
        get:
            tags:
                - Registry
            description: List all applications
            operationId: Registry_ListApps
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.registry.v1.ListAppsReply'
        post:
            tags:
                - Registry
            description: Create a new application
            operationId: Registry_CreateApp
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.registry.v1.CreateAppRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.registry.v1.CreateAppReply'
    /v1/apps/{appId}:
        get:
            tags:
                - Registry
            description: Get application details
            operationId: Registry_GetApp
            parameters:
                - name: appId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.registry.v1.GetAppReply'
    /v1/apps/{appId}/keys:
        post:
            tags:
                - Registry
            description: Issue a new API key for an application
            operationId: Registry_IssueApiKey
            parameters:
                - name: appId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.registry.v1.IssueApiKeyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.registry.v1.IssueApiKeyReply'
...
    /v1/keys/{keyHash}:
        delete:
            tags:
                - Registry
            description: Revoke an existing API key
            operationId: Registry_RevokeApiKey
            parameters:
                - name: keyHash
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.registry.v1.RevokeApiKeyReply'
...
    /v1/rules:
        get:
            tags:
                - Rules
            description: List all rules for the app
            operationId: Rules_ListRules
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rules.v1.ListRulesReply'
        post:
            tags:
                - Rules
            description: Create a new business rule
            operationId: Rules_CreateRule
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rules.v1.CreateRuleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rules.v1.RuleReply'
    /v1/rules/{id}:
        get:
            tags:
                - Rules
            description: Get a specific rule
            operationId: Rules_GetRule
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rules.v1.RuleReply'
components:
    schemas:
...
        api.registry.v1.CreateAppReply:
            type: object
            properties:
                appId:
                    type: string
                status:
                    type: string
        api.registry.v1.CreateAppRequest:
            type: object
            properties:
                appName:
                    type: string
                description:
                    type: string
                owner:
                    type: string
        api.registry.v1.GetAppReply:
            type: object
            properties:
                appId:
                    type: string
                appName:
                    type: string
                description:
                    type: string
                owner:
                    type: string
                status:
                    type: string
        api.registry.v1.IssueApiKeyReply:
            type: object
            properties:
                apiKey:
                    type: string
                keyHash:
                    type: string
                keyPrefix:
                    type: string
        api.registry.v1.IssueApiKeyRequest:
            type: object
            properties:
                appId:
                    type: string
                name:
                    type: string
                scopes:
                    type: string
                ttlSeconds:
                    type: string
        api.registry.v1.ListAppsReply:
            type: object
            properties:
                apps:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.registry.v1.GetAppReply'
        api.registry.v1.RevokeApiKeyReply:
            type: object
            properties:
                success:
                    type: boolean
        api.rules.v1.CreateRuleRequest:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                triggerType:
                    type: string
                cron:
                    type: string
                cypherQuery:
                    type: string
                action:
                    type: string
                payloadJson:
                    type: string
        api.rules.v1.ListRulesReply:
            type: object
            properties:
                rules:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rules.v1.RuleReply'
        api.rules.v1.RuleReply:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                description:
                    type: string
                triggerType:
                    type: string
                cron:
                    type: string
                cypherQuery:
                    type: string
                action:
                    type: string
                payloadJson:
                    type: string
                isActive:
                    type: boolean
...